// Frontend/src/lib/types/typeheavyMachineryRegistration.tsx
// lib/submitVehicleForm.ts
export interface heavyMachineryRegistration {  
    listingTitle: string;
          make: string;
          model: string;
          condition: string;
          inventoryLocation: string;
          transmission: string;
          fuelType: string;
          axleConfig: string;
          hotLocation: string;
          steering: string;
          color: string;
          category: string;
          price: number;
          operationalHours: number;
          year: number;
          dealerStock: boolean;
          userId: string;
          imageBase64: string;
          imageUrl?: string;
          carOptions: string[];
          description: string;
          adminNotes: string;
  }
  
    